Fix negative binomial implementation
Use regular C++ support to implement it instead of pretending to be
smart.

+++ pythran/pythonic/include/numpy/random/negative_binomial.hpp
@@ -1,10 +1,10 @@
 #ifndef PYTHONIC_INCLUDE_NUMPY_RANDOM_NEGATIVE_BINOMIAL_HPP
 #define PYTHONIC_INCLUDE_NUMPY_RANDOM_NEGATIVE_BINOMIAL_HPP
 
-#include "pythonic/include/utils/functor.hpp"
-#include "pythonic/include/types/ndarray.hpp"
 #include "pythonic/include/types/NoneType.hpp"
+#include "pythonic/include/types/ndarray.hpp"
 #include "pythonic/include/types/tuple.hpp"
+#include "pythonic/include/utils/functor.hpp"
 
 PYTHONIC_NS_BEGIN
 namespace numpy
@@ -12,17 +12,17 @@ namespace numpy
   namespace random
   {
     template <class pS>
-    types::ndarray<double, pS> negative_binomial(double n, double p,
-                                                 pS const &shape);
+    types::ndarray<long, pS> negative_binomial(long n, double p,
+                                               pS const &shape);
 
-    auto negative_binomial(double n, double p, long size)
+    auto negative_binomial(long n, double p, long size)
         -> decltype(negative_binomial(n, p, types::array<long, 1>{{size}}));
 
-    double negative_binomial(double n, double p, types::none_type size = {});
+    long negative_binomial(long n, double p, types::none_type size = {});
 
     DEFINE_FUNCTOR(pythonic::numpy::random, negative_binomial);
-  }
-}
+  } // namespace random
+} // namespace numpy
 PYTHONIC_NS_END
 
 #endif
